backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-owasp-zap-v2.4==0.0.20
sqlalchemy==2.0.25
requests==2.31.0
google-generativeai
